feat: create usePokemons hook and refactor homepage
