+++ class-components/src/pages/HomePage.tsx
@@ -1,4 +1,3 @@
-import { useState, useEffect, useCallback } from 'react';
 import {
   useSearchParams,
   useNavigate,
@@ -7,74 +6,22 @@ import {
   useOutlet,
 } from 'react-router-dom';
 import { useLocalStorage } from '../hooks/useLocalStorage';
-import { getPokemonList, getPokemonFullDetails } from '../api/pokemonService';
-
+import { usePokemons } from '../hooks/usePokemons';
 import Main from '../components/Main/Main';
 import Pagination from '../components/Pagination/Pagination';
 import Header from '../components/Header/Header';
-import type { PokemonDetails } from '../types';
-
-const POKEMON_PER_PAGE = 20;
 
 const HomePage = () => {
-  const [pokemons, setPokemons] = useState<PokemonDetails[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState<Error | null>(null);
-  const [nextPageUrl, setNextPageUrl] = useState<string | null>(null);
-  const [prevPageUrl, setPrevPageUrl] = useState<string | null>(null);
-  const [totalPages, setTotalPages] = useState(0);
-
   const [searchTerm, setSearchTerm] = useLocalStorage('searchTerm', '');
   const [searchParams, setSearchParams] = useSearchParams();
-
   const navigate = useNavigate();
   const location = useLocation();
   const outlet = useOutlet();
 
   const currentPage = parseInt(searchParams.get('page') || '1', 10);
 
-  const fetchPokemonsByUrl = useCallback(async (url: string) => {
-    setIsLoading(true);
-    setError(null);
-    try {
-      const { pokemons, next, previous, count } = await getPokemonList(url);
-      setPokemons(pokemons);
-      setNextPageUrl(next);
-      setPrevPageUrl(previous);
-      setTotalPages(Math.ceil(count / POKEMON_PER_PAGE));
-    } catch (err) {
-      setError(err as Error);
-    } finally {
-      setIsLoading(false);
-    }
-  }, []);
-
-  const fetchSinglePokemon = useCallback(async (name: string) => {
-    setIsLoading(true);
-    setError(null);
-    setNextPageUrl(null);
-    setPrevPageUrl(null);
-    try {
-      const pokemon = await getPokemonFullDetails(name.toLowerCase());
-      setPokemons([pokemon]);
-      setTotalPages(1);
-    } catch (err) {
-      setError(new Error(`Pokemon "${name}" not found.`));
-      setPokemons([]);
-    } finally {
-      setIsLoading(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    if (searchTerm) {
-      fetchSinglePokemon(searchTerm);
-    } else {
-      const offset = (currentPage - 1) * POKEMON_PER_PAGE;
-      const url = `https://pokeapi.co/api/v2/pokemon?limit=${POKEMON_PER_PAGE}&offset=${offset}`;
-      fetchPokemonsByUrl(url);
-    }
-  }, [currentPage, searchTerm, fetchPokemonsByUrl, fetchSinglePokemon]);
+  const { pokemons, isLoading, error, nextPageUrl, prevPageUrl, totalPages } =
+    usePokemons(currentPage, searchTerm);
 
   const handleSearch = (term: string) => {
     setSearchTerm(term);
@@ -101,7 +48,7 @@ const HomePage = () => {
     navigate(`/${location.search}`);
   };
 
-  const showPagination = !searchTerm && !error;
+  const showPagination = !searchTerm && !error && pokemons.length > 1;
 
   return (
     <>
